requirements.txt
beautifulsoup4==4.12.3
Pillow==10.3.0
reportlab==4.2.0
Requests==2.31.0
selenium==4.20.0
